Log dev

+++ platform/native/ios-xcode/tt_semaphore_native.c
@@ -1,4 +1,6 @@
-/* Licensed to the Apache Software Foundation (ASF) under one or more
+/* Copyright (C) 2017 haniu (niuhao.cn@gmail.com)
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
  * this work for additional information regarding copyright ownership.
  * The ASF licenses this file to You under the Apache License, Version 2.0
@@ -91,7 +93,7 @@ tt_bool_t tt_sem_acquire_ntv(IN tt_sem_ntv_t *sys_sem, IN tt_s64_t wait_ms)
         mach_timespec_t t;
         kern_return_t r;
 
-        t.tv_sec = wait_ms / 1000;
+        t.tv_sec = (unsigned int)(wait_ms / 1000);
         t.tv_nsec = (wait_ms % 1000) * 1000000;
 
         // again2:
